v1.0.5
- Before the same code was applied for each move of the mouse... that is
really highly consuming ! Now we just create the image during the
initialization phase, and that's it.
- Before we were forced to use a top element instead of choosing the
elements we want to apply the function. Now we can directly define the
elements concerned by the detection. Eg:
$('#my_element').detectmouseoverdisabled() .... And for compabitlity
with the previous versions we can still do
$('body').detectmouseoverdisabled() that will apply the function to all
input, textarea and select elements that are disabled
- Add some comments into the code
- Tested with last Chrome and Firefox 14, also with IE8
- When an element is hidden then jQuery returns a wrong width... so
using getWidth() function to get the real width
- Use markdown for the README and adding some basic text
- Rename the version to 1.0.5